moved render and date creation functions to utils
